# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get all videos uploaded by the authenticated user",
        "operationId": "get_user_videos",
        "responses": {
          "200": {
            "description": "List of videos uploaded by the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/videos/upload": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Handle video upload",
        "operationId": "upload_video",
        "requestBody": {
          "content": {
            "multipart/form-data": {}
          }
        },
        "responses": {
          "200": {
            "description": "Video uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing video file or invalid data"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error - failed to save file or database error"
          }
        }
      }
    },
    "/api/videos/{id}/stream": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Stream a video file with Range request support",
        "operationId": "stream_video",
        "responses": {
          "200": {
            "description": "Full video file"
          },
          "206": {
            "description": "Partial content (range request)"
          },
          "404": {
            "description": "Video not found"
          },
          "416": {
            "description": "Range not satisfiable"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UploadResponse": {
        "type": "object",
        "required": [
          "id",
          "message"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Video": {
        "type": "object",
        "required": [
          "id",
          "file_path",
          "original_filename",
          "user_id",
          "uploaded_at"
        ],
        "properties": {
          "file_path": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "original_filename": {
            "type": "string"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string"
          }
        }
      }
    }
  }
}